requirements.txt
psycopg2
sqlalchemy
pandas
torch
python-decouple==3.8
pyspark
